refactor: centralize sign-in schema and update import paths
- Moved the sign-in schema to a new types directory for better organization and reusability.
- Updated import paths in authentication-related files to reference the new schema location.
- Cleaned up the sign-in action and form components to reflect the new structure, ensuring functionality remains intact.
